requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=0.25.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
blobfile=2.1.1=pypi_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
diffusers=0.24.0=pypi_0
docker-pycreds=0.4.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=pypi_0
fonttools=4.47.0=pypi_0
fsspec=2023.12.2=pypi_0
ftfy=6.1.3=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
huggingface-hub=0.19.4=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.0.0=pyha770c72_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=7.0.0=hd8ed1ab_0
ipykernel=6.26.0=pyhf8b6a83_0
ipython=8.18.1=pyh707e725_3
ithaca365=1.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.5.1=py39hf3d152e_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_3
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=13.2.0=h807b86a_3
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
lxml=4.9.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpi=1.0=mpich
mpi4py=3.1.4=py39hfc96bbd_0
mpich=3.3.2=hc856adb_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
ninja=1.11.1.1=pypi_0
numpy=1.26.2=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.2.0=hd590300_1
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.1.0=pypi_0
pip=23.3.1=py39h06a4308_0
platformdirs=4.1.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.25.1=pypi_0
psutil=5.9.7=py39hd1e30aa_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-cpuinfo=9.0.0=pypi_0
pycocotools=2.0.7=pypi_0
pycryptodomex=3.19.0=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
pyquaternion=0.9.9=pypi_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch-fid=0.3.0=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=23.0.0=py39headdf64_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
safetensors=0.4.1=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
seaborn=0.13.1=pypi_0
sentry-sdk=1.39.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.2.0=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.1.2=pypi_0
torchaudio=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.3.3=py39hd1e30aa_1
tqdm=4.66.1=pypi_0
traitlets=5.14.0=pyhd8ed1ab_0
triton=2.1.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.3=pypi_0
ultralytics=8.1.5=pypi_0
urllib3=2.1.0=pypi_0
wandb=0.16.1=pypi_0
wcwidth=0.2.12=pyhd8ed1ab_0
wheel=0.41.2=py39h06a4308_0
xformers=0.0.23.post1=pypi_0
xz=5.4.5=h5eee18b_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
